W136 trim FUN_003f9510 widths

+++ src/Event/Fcl/fclShopMisc.c
@@ -302,7 +302,7 @@ void FUN_003f86a0(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f8a00(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f8e10(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f9220(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5 );
-void FUN_003f9510(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5 );
+void FUN_003f9510(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 u32 FUN_003f98f0(u16 param_1);
 u64 FUN_003f99d0(u32 param_1,u32 param_2);
 u64 FUN_003f9b20(u64 param_1,int param_2);
@@ -7476,7 +7476,7 @@ void FUN_003f9220(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5)
 // FUN_003F9510 NONMATCHING
 
 
-void FUN_003f9510(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5
+void FUN_003f9510(int param_1,int param_2,u32 param_3,int param_4,int param_5
 
                  )
 
@@ -7488,7 +7488,7 @@ void FUN_003f9510(u64 param_1,u64 param_2,u32 param_3,u64 param_4,int param_5
 
   u32 *puVar2;
 
-  long lVar3;
+  int lVar3;
 
   int iVar4;
